requirements.txt
altgraph==0.10.2
appdirs==1.4.3
asn1crypto==0.24.0
astroid==1.6.6
backports.functools-lru-cache==1.6.1
bdist-mpkg==0.5.0
# bonjour-py==0.3
certifi==2020.4.5.1
cffi==1.12.2
chardet==3.0.4
configparser==4.0.2
contextlib2==0.6.0.post1
cryptography==2.6.1
distlib==0.3.0
django-crispy-forms==1.9.0
enum34==1.1.6
filelock==3.0.12
future==0.17.1
# futures==3.3.0
idna==2.9
importlib-metadata==1.6.0
importlib-resources==1.5.0
ipaddress==1.0.22
isort==4.3.21
lazy-object-proxy==1.4.3
macholib==1.5.1
# matplotlib==1.3.1
mccabe==0.6.1
modulegraph==0.10.4
numpy==1.16.6
opencv-contrib-python==3.4.2.17
opencv-python==3.4.2.17
pathlib2==2.3.5
py2app==0.7.3
pycparser==2.19
pygame==2.0.0.dev4
# pylint==1.9.5
pyobjc-core==2.5.1
pyobjc-framework-Accounts==2.5.1
pyobjc-framework-AddressBook==2.5.1
pyobjc-framework-AppleScriptKit==2.5.1
pyobjc-framework-AppleScriptObjC==2.5.1
pyobjc-framework-Automator==2.5.1
pyobjc-framework-CFNetwork==2.5.1
pyobjc-framework-Cocoa==2.5.1
pyobjc-framework-Collaboration==2.5.1
pyobjc-framework-CoreData==2.5.1
pyobjc-framework-CoreLocation==2.5.1
pyobjc-framework-CoreText==2.5.1
pyobjc-framework-DictionaryServices==2.5.1
pyobjc-framework-EventKit==2.5.1
pyobjc-framework-ExceptionHandling==2.5.1
pyobjc-framework-FSEvents==2.5.1
pyobjc-framework-InputMethodKit==2.5.1
pyobjc-framework-InstallerPlugins==2.5.1
pyobjc-framework-InstantMessage==2.5.1
pyobjc-framework-LatentSemanticMapping==2.5.1
pyobjc-framework-LaunchServices==2.5.1
pyobjc-framework-Message==2.5.1
pyobjc-framework-OpenDirectory==2.5.1
pyobjc-framework-PreferencePanes==2.5.1
pyobjc-framework-PubSub==2.5.1
pyobjc-framework-QTKit==2.5.1
pyobjc-framework-Quartz==2.5.1
pyobjc-framework-ScreenSaver==2.5.1
pyobjc-framework-ScriptingBridge==2.5.1
pyobjc-framework-SearchKit==2.5.1
pyobjc-framework-ServiceManagement==2.5.1
pyobjc-framework-Social==2.5.1
pyobjc-framework-SyncServices==2.5.1
pyobjc-framework-SystemConfiguration==2.5.1
pyobjc-framework-WebKit==2.5.1
pyOpenSSL==19.0.0
pyparsing==2.0.1
python-dateutil==1.5
pytz==2013.7
requests==2.23.0
scandir==1.10.0
# scipy==0.13.0b1
singledispatch==3.4.0.3
six==1.12.0
typing==3.7.4.1
urllib3==1.25.8
virtualenv==20.0.18
virtualenvwrapper-win==1.2.6
wrapt==1.12.1
# xattr==0.6.4
zipp==1.2.0
